pyproject.toml
[tool.poetry]
name = "Nohara calendar"
version = "0.1.0"
description = "野原工芸の来店予約カレンダーを監視し，空きが発生した場合に通知を行うスクリプトです．"
authors = ["KIMATA Tetsuya <kimata@green-rabbit.net>"]

packages = [
    { include = "local_lib", from = "lib" },
    { include = "calendar_check", from = "lib" }
]

[tool.poetry.dependencies]
coloredlogs = "^15.0.1"
docopt = "^0.6.2"
python = "^3.10"
pyyaml = "^6.0.1"
selenium = "^4.18.1"
pyprind = "^2.11.3"
enlighten = "^1.12.4"
openpyxl = "^3.1.2"
pillow = "^10.2.0"
imageio = "^2.34.0"
jinxed = "^1.2.1"
line-bot-sdk = "^3.11.0"

[tool.poetry.group.dev.dependencies]
nuitka = "^2.1.3"

[tool.pytest.ini_options]
minversion = "6.0"

addopts = "--verbose --cov=app --cov=lib --log-file-level=INFO --log-format=\"%(asctime)s %(levelname)s %(message)s\" --log-format=\"%(asctime)s %(levelname)s [%(filename)s:%(lineno)s %(funcName)s] %(message)s\" --capture=sys --cov-report=html --html=tests/evidence/index.htm --self-contained-html"

testpaths = [
    "tests",
]
filterwarnings = [
    "ignore:The hookimpl CovPlugin.pytest_configure_node uses",
    "ignore:The hookimpl CovPlugin.pytest_testnodedown uses",
    "ignore::DeprecationWarning:pytest_freezegun",
]

[tool.coverage.run]
branch = true

[tool.coverage.report]
exclude_lines = [
  "pragma: no cover",
  "if __name__ == .__main__.:",

  # 以下，未使用関数
]

[tool.coverage.html]
directory = "tests/evidence/coverage"

[tool.black]
target-version = ['py310']
line-length = 110

[tool.isort]
profile = "black"
line_length = 110
include_trailing_comma = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
